rebased onto last and ready to merge
